feat: use new component everywhere applicable
#83
